+++ pool/web/view/component/component_contacts.ml
@@ -41,3 +41,53 @@ let identity_with_icons view_contact_name contact entity_id =
   span [ txt text ] :: status_icons contact
   |> div ~a:[ a_class [ "flexrow"; "flex-gap-sm" ] ]
 ;;
+
+let external_data_ids { Pool_context.language; _ } external_data_ids =
+  match external_data_ids with
+  | [] ->
+    p [ txt Pool_common.(Utils.text_to_string language I18n.EmptyListGeneric) ]
+  | external_data_ids ->
+    let thead =
+      let open Pool_common in
+      Message.Field.[ Experiment; Session; ExternalDataId ]
+      |> CCList.map CCFun.(Utils.field_to_string_capitalized language %> txt)
+    in
+    external_data_ids
+    |> CCList.map
+         (fun
+             Assignment.
+               { ExternalDataIdentifier.external_data_id
+               ; experiment_id
+               ; experiment_title
+               ; session_id
+               ; session_start
+               ; session_duration
+               }
+           ->
+            let experiment_path =
+              Format.asprintf
+                "/admin/experiments/%s"
+                Experiment.(Id.value experiment_id)
+            in
+            [ a
+                ~a:[ a_href (experiment_path |> Sihl.Web.externalize_path) ]
+                [ txt (Experiment.Title.value experiment_title) ]
+            ; a
+                ~a:
+                  [ a_href
+                      (Format.asprintf
+                         "%s/sessions/%s"
+                         experiment_path
+                         (Session.Id.value session_id)
+                       |> Sihl.Web.externalize_path)
+                  ]
+                [ txt
+                    Session.(
+                      Utils.Ptime.format_datetime_with_span
+                        (Start.value session_start)
+                        (Duration.value session_duration))
+                ]
+            ; txt Assignment.(ExternalDataId.value external_data_id)
+            ])
+    |> Component_table.horizontal_table ~thead `Striped
+;;