+++ pool/web/view/page/page_admin_contact.ml
@@ -268,58 +268,12 @@ let edit
     ]
 ;;
 
-let external_data_ids { Pool_context.language; _ } contact external_data_ids =
-  let open Assignment in
-  let table =
-    match external_data_ids with
-    | [] ->
-      p
-        [ txt Pool_common.(Utils.text_to_string language I18n.EmptyListGeneric)
-        ]
-    | external_data_ids ->
-      let thead =
-        let open Pool_common in
-        Message.Field.[ Experiment; Session; ExternalDataId ]
-        |> CCList.map CCFun.(Utils.field_to_string_capitalized language %> txt)
-      in
-      external_data_ids
-      |> CCList.map
-           (fun
-               { ExternalDataIdentifier.external_data_id
-               ; experiment_id
-               ; experiment_title
-               ; session_id
-               ; session_start
-               ; session_duration
-               }
-             ->
-              let experiment_path =
-                Format.asprintf
-                  "/admin/experiments/%s"
-                  Experiment.(Id.value experiment_id)
-              in
-              [ a
-                  ~a:[ a_href (experiment_path |> Sihl.Web.externalize_path) ]
-                  [ txt (Experiment.Title.value experiment_title) ]
-              ; a
-                  ~a:
-                    [ a_href
-                        (Format.asprintf
-                           "%s/sessions/%s"
-                           experiment_path
-                           (Session.Id.value session_id)
-                         |> Sihl.Web.externalize_path)
-                    ]
-                  [ txt
-                      Session.(
-                        Utils.Ptime.format_datetime_with_span
-                          (Start.value session_start)
-                          (Duration.value session_duration))
-                  ]
-              ; txt Assignment.(ExternalDataId.value external_data_id)
-              ])
-      |> Component.Table.horizontal_table ~thead `Striped
-  in
+let external_data_ids
+  ({ Pool_context.language; _ } as context)
+  contact
+  external_data_ids
+  =
+  let table = Component.Contacts.external_data_ids context external_data_ids in
   div
     ~a:[ a_class [ "trim"; "safety-margin"; "stack" ] ]
     [ h1 [ txt (Contact.fullname contact) ]